Attemping POST for SelectedShop
