+++ src/components/App/App.js
@@ -4,17 +4,37 @@ import Header from "../Header/Header";
 import CardContainer from "../CardContainer/CardContainer";
 import SelectedShop from "../SelectedShop/SelectedShop";
 import sampleCoffeeShops from "../../MockData";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { Routes, Route, Link, useLocation } from "react-router-dom";
 
 function App() {
-  const [shops, setShops] = useState(sampleCoffeeShops);
+  const [shops, setShops] = useState([]);
+
+  const getShops = () => {
+    return fetch(`http://localhost:3001/`)
+      .then(response => {
+        if(!response.ok) {
+          throw new Error(`${response.status} ${response.statusText}`)
+        }
+        console.log('hello', response.json)
+        return response.json()
+      })
+  }
+  
+  useEffect(() => {
+    getShops()
+      .then(data => setShops(data))
+      .catch(error => console.log(error.message))
+  }, [])
+  
+
+
 
   return (
     <div className='App'>
       <Header />
       <Routes>
-        <Route path='/' element={<CardContainer />} />
+        <Route path='/' element={<CardContainer shops={shops}/>} />
         <Route path='/SelectedShop/:id' element={<SelectedShop shops={shops} />} />
         <Route path='/*' element={<Error />} />
       </Routes>
